Gameboard dots

+++ game/MP-Client-Test/src/org/golde/saas/risingseasgame/client/objects/GameObjectMoveable.java
@@ -2,25 +2,25 @@ package org.golde.saas.risingseasgame.client.objects;
 
 public abstract class GameObjectMoveable implements GameObject {
 
-	private float posX, posY;
+	private int posX, posY;
 
-	public final float getX() {
+	public final int getX() {
 		return posX;
 	}
 	
-	public final float getY() {
+	public final int getY() {
 		return posY;
 	}
 	
-	public final void setX(float x) {
+	public final void setX(int x) {
 		this.posX = x;
 	}
 	
-	public final void setY(float y) {
+	public final void setY(int y) {
 		this.posY = y;
 	}
 	
-	public final void setXY(float x, float y) {
+	public final void setXY(int x, int y) {
 		setX(x);
 		setY(y);
 	}
